chore: remove generic git from detector (#24)
This commit removes generic git from the list of build type detectors. The reason is that we don't want to download anything big there until we solve the resources issue. So for the unknown repositories or for the cases when ssh key is used, the detector will respond just empty statistics.